requirements.txt
streamlit
pandas
pymupdf
google-generativeai
pillow
openpyxl